+++ src/main/java/com/yyx/mapr/test/WordCountReducer.java
@@ -1,21 +1,22 @@
 package com.yyx.mapr.test;
 
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
 import java.io.IOException;
 import java.util.Iterator;
 
-public class WordCountReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+public class WordCountReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
 
     @Override
-    protected void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
+    protected void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
         int sum = 0;
-        for (IntWritable i : values
+        for (LongWritable i : values
                 ) {
             sum += i.get();
         }
-        context.write(key, new IntWritable(sum));
+        context.write(key, new LongWritable(sum));
     }
 }
